requirements.txt
adal==1.2.7
autopep8==2.0.1
azure-common==1.1.28
azure-core==1.26.1
azure-functions==1.12.0
azure-identity==1.12.0
azure-keyvault-secrets==4.6.0
azure-mgmt-containerinstance==10.0.0
azure-mgmt-core==1.3.2
azure-mgmt-resource==22.0.0
azure-storage-blob==12.14.1
azure-storage-file-share==12.10.1
azure-storage-queue==12.5.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.0
dnspython==2.2.1
idna==3.4
isodate==0.6.1
msal==1.20.0
msal-extensions==1.0.0
msrest==0.7.1
msrestazure==0.6.4
oauthlib==3.2.2
portalocker==2.6.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.4
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
tomli==2.0.1
typing-extensions==4.4.0
urllib3==1.26.13
